requirements.txt
requests==2.28.0
beautifulsoup4==4.11.1
PyPDF2==2.10.0
tiktoken==0.6.0
nltk==3.7
nbformat==5.4.0
nbconvert==6.5.0
youtube-transcript-api==0.4.1
pyperclip==1.8.2
wget==3.2
tqdm==4.64.0
rich==12.4.4
Flask==3.1.0
PyYAML==6.0.1
